+++ pallets/protos/src/mock.rs
@@ -13,7 +13,19 @@ use sp_runtime::{
 
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
 type Block = frame_system::mocking::MockBlock<Test>;
-
+pub const DATA: &str = "0x0155a0e40220";
+pub const PROTO_HASH: Hash256 = [
+	30, 138, 136, 186, 232, 46, 112, 65, 122, 54, 110, 89, 123, 195, 7, 150, 12, 134, 10, 179, 245,
+	51, 83, 227, 72, 251, 5, 148, 207, 251, 119, 59,
+];
+pub const PUBLIC: [u8; 33] = [
+	3, 137, 65, 23, 149, 81, 74, 241, 98, 119, 101, 236, 239, 252, 189, 0, 39, 25, 240, 49, 96, 79,
+	173, 215, 209, 136, 226, 220, 88, 91, 78, 26, 251,
+];
+pub const PUBLIC1: [u8; 32] = [
+	137, 65, 23, 149, 81, 74, 241, 98, 119, 101, 236, 239, 252, 189, 0, 39, 25, 240, 49, 96, 79,
+	173, 215, 209, 136, 226, 220, 88, 91, 78, 26, 251,
+];
 
 // Configure a mock runtime to test the pallet.
 frame_support::construct_runtime!(
